+++ packages/storybook/src/utilities/states.ts
@@ -58,6 +58,12 @@ export const requiredVisibleStates = [
 ] as const;
 export type RequiredVisibleState = (typeof requiredVisibleStates)[number];
 
+export const fullBleedStates = [
+    ['', false],
+    ['Full Bleed', true]
+] as const;
+export type FullBleedState = (typeof fullBleedStates)[number];
+
 export const disabledReadOnlyStates = [
     ['', false, false, false],
     ['Appearance-Read-Only', false, false, true],
@@ -92,3 +98,15 @@ export const onlyDisabledAbsentStates = disabledReadOnlyStates.filter(
     ) => state[2] === false
 );
 export type OnlyDisabledAbsentState = (typeof onlyDisabledAbsentStates)[number];
+
+export const onlyReadOnlyAbsentStates = disabledReadOnlyStates.filter(
+    (
+        state: readonly [
+            name: string,
+            readOnly: boolean,
+            disabled: boolean,
+            appearanceReadonly: boolean
+        ]
+    ) => state[1] === false
+);
+export type OnlyReadOnlyAbsentState = (typeof onlyReadOnlyAbsentStates)[number];