+++ packages/angular-workspace/nimble-angular/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @ni/nimble-angular
 
-<!-- This log was last generated on Wed, 14 May 2025 18:31:13 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 20 May 2025 23:06:06 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 30.2.0
+
+Tue, 20 May 2025 23:06:06 GMT
+
+### Minor changes
+
+- Bump @ni/nimble-components to v33.7.0
+
+## 30.1.6
+
+Tue, 20 May 2025 21:33:37 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v33.6.0
+
+## 30.1.5
+
+Tue, 20 May 2025 21:01:44 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v33.5.0
 
 ## 30.1.4